+++ dexpaprika_sdk/__init__.py
@@ -27,7 +27,7 @@ from .models import (
     Stats
 )
 
-__version__ = "0.6.0"
+__version__ = "0.7.0"
 __all__ = [
     "DexPaprikaClient",
     # Exceptions